SB-536: Implement cron tasks
Extracted an actual interface for the CronTaskConfigurationService.
